added fixed; leevel generator

+++ backend/app/models.py
@@ -27,7 +27,7 @@ class GetLevelMongo(BaseModel):
     level_id: int
     prices: List[float]
     tones: List[float]
-    volumes: List[int]
+    volumes: List[float]
     news: List[Tuple[str, str]]
     model_predict: int
     target: int
@@ -39,7 +39,7 @@ class GetLevelResponse(BaseModel):
     dates: List[str]
     prices: List[float]
     tones: List[float]
-    volumes: List[int]
+    volumes: List[float]
     news: List[Tuple[str, str]]
     model_predict: int
     target: int
