+++ frontend/app/models.py
@@ -7,7 +7,7 @@ class Level(BaseModel):
     dates: List[str]
     prices: List[float]
     tones: List[float]
-    volumes: List[int]
+    volumes: List[float]
     news: List[Tuple[str, str]]
     model_predict: int
     target: int
